+++ src/client/types/swap.ts
@@ -1,11 +1,13 @@
 export type SwapSimulationResponse = {
   ask_address: string;
+  ask_jetton_wallet: string;
   ask_units: string;
   fee_address: string;
   fee_percent: string;
   fee_units: string;
   min_ask_units: string;
   offer_address: string;
+  offer_jetton_wallet: string;
   offer_units: string;
   pool_address: string;
   price_impact: string;